Zenbook Event Scheduler
A event scheduler is a timetable for an event , program to held along with its details. The event schedule generates an event once at the specified date and time. The event schedule generates an event on a regular basis, You can also configure the event schedule to generate an event multiple times on the specified day.